Using ts.factory.createTypeAliasDeclaration leads to the creation of a node that doesn't have a source file attached to it. In some circumstances this will result in invalid code.
For example `export type Something = "a" | "b"` would turn into: `type Something = | `.

This is because printNode() will create an empty source file if the node to be printed
doesn't have a source file attached to it. My guess is that TypeScript associates string literals
with the source file to save space.

Because of this, we will treat the removal of the export modifier and renaming of the type as a
source file transformation.